requirements.txt
fastapi
uvicorn
sqlalchemy
Pillow
numpy
python-multipart